requirements.txt
psycopg2>=2.9.9
PyYAML>=6.0.1
semver>=3.0.2
python-dotenv>=1.0.1
pycryptodome>=3.20.0